universal_adk_agent/requirements.txt
google-generativeai
google-cloud-aiplatform
streamlit
python-dotenv
pandas
yfinance
duckduckgo-search
langchain
langgraph
# Assuming google-adk is part of the environment or available via pip
# google-adk
